src/requirements.txt
chainlit
openai-agents>=0.0.19
python-dotenv>=1.0.0
requests>=2.31.0
